Directory structure changes and README.md fixes/additions.
This commit makes small changes to the structure of the powerskill,
where previously the function ran straight from __init__.py, now it
runs from a separate main.py. This is done to facilitate testing and
cross-module imports. Also, small fixes to README.md were added as
well as some documentation related to testing.